Refactor home page to redirect to create, add dedicated feed page
